+++ src/main/java/racingcar/RacingGame.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import java.util.List;
 import racingcar.domain.MoveResults;
+import racingcar.domain.RacingCircuit;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -9,12 +10,12 @@ public class RacingGame {
 
     private final InputView inputView;
     private final OutputView outputView;
-    private final Circuit circuit;
+    private final RacingCircuit racingCircuit;
 
-    public RacingGame(final InputView inputView, final OutputView outputView, final Circuit circuit) {
+    public RacingGame(final InputView inputView, final OutputView outputView, final RacingCircuit racingCircuit) {
         this.inputView = inputView;
         this.outputView = outputView;
-        this.circuit = circuit;
+        this.racingCircuit = racingCircuit;
     }
 
     public void run() {
@@ -25,17 +26,17 @@ public class RacingGame {
 
     private void registerCars() {
         List<String> carNames = inputView.inputCarNames();
-        circuit.registerCars(carNames);
+        racingCircuit.registerCars(carNames);
     }
 
     private void startRace() {
         int raceTryCount = inputView.inputRaceTryCount();
-        List<MoveResults> moveResults = circuit.startRace(raceTryCount);
+        List<MoveResults> moveResults = racingCircuit.startRace(raceTryCount);
         outputView.printMoveResults(moveResults);
     }
 
     private void awardRace() {
-        List<String> winners = circuit.findWinners();
+        List<String> winners = racingCircuit.findWinners();
         outputView.printWinners(winners);
     }
 }
